Vivendo Perigosamente

+++ api/src/main/java/br/pro/fernandocesar/basetcc/api/controller/FotoOcorrenciaController.java
@@ -1,6 +1,7 @@
 package br.pro.fernandocesar.basetcc.api.controller;
 
 import br.pro.fernandocesar.basetcc.api.domain.FotoOcorrencia;
+import br.pro.fernandocesar.basetcc.api.domain.Ocorrencia;
 import br.pro.fernandocesar.basetcc.api.service.FotoOcorrenciaService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
@@ -28,7 +29,7 @@ public class FotoOcorrenciaController {
     }
 
     @GetMapping("/ocorrencia/{ocorrencia}")
-    public List<FotoOcorrencia> getByOcorrencia(@PathVariable int ocorrencia) {
+    public List<FotoOcorrencia> getByOcorrencia(@PathVariable long ocorrencia) {
         return service.findByOcorrencia(ocorrencia);
     }
